Decoule print preview extraction logic from AIChat* classes into PrintPreviewExtractor
