+++ browser/ui/webui/ai_chat/ai_chat_ui_page_handler.cc
@@ -11,24 +11,18 @@
 #include <utility>
 #include <vector>
 
-#include "base/json/json_reader.h"
 #include "base/notreached.h"
 #include "base/strings/utf_string_conversions.h"
 #include "brave/browser/ui/side_panel/ai_chat/ai_chat_side_panel_utils.h"
-#include "brave/browser/ui/webui/ai_chat/ai_chat_ui.h"
 #include "brave/components/ai_chat/core/browser/constants.h"
 #include "brave/components/ai_chat/core/browser/models.h"
-#include "brave/components/ai_chat/core/browser/utils.h"
 #include "brave/components/ai_chat/core/common/mojom/ai_chat.mojom-shared.h"
 #include "brave/components/ai_chat/core/common/mojom/ai_chat.mojom.h"
 #include "brave/components/ai_chat/core/common/pref_names.h"
-#include "brave/components/text_recognition/common/buildflags/buildflags.h"
 #include "chrome/browser/favicon/favicon_service_factory.h"
-#include "chrome/browser/pdf/pdf_pref_names.h"
 #include "chrome/browser/profiles/profile.h"
 #include "chrome/browser/ui/browser_finder.h"
 #include "chrome/browser/ui/singleton_tabs.h"
-#include "chrome/common/pref_names.h"
 #include "components/favicon/core/favicon_service.h"
 #include "components/prefs/pref_service.h"
 #include "content/public/browser/browser_context.h"
@@ -38,17 +32,6 @@
 #include "content/public/browser/web_contents_observer.h"
 #include "content/public/common/url_constants.h"
 #include "ui/base/l10n/l10n_util.h"
-
-#if BUILDFLAG(ENABLE_PRINT_PREVIEW)
-#include "brave/services/printing/public/mojom/pdf_to_bitmap_converter.mojom.h"
-#include "chrome/browser/printing/print_preview_data_service.h"
-#include "chrome/browser/printing/print_view_manager_common.h"
-#include "chrome/browser/printing/printing_service.h"
-#include "chrome/services/printing/public/mojom/printing_service.mojom.h"
-#include "printing/print_job_constants.h"
-#include "third_party/blink/public/common/associated_interfaces/associated_interface_provider.h"
-#include "third_party/skia/include/core/SkBitmap.h"
-#endif
 
 #if BUILDFLAG(IS_ANDROID)
 #include "brave/browser/ui/android/ai_chat/brave_leo_settings_launcher_helper.h"
@@ -71,115 +54,12 @@ using mojom::CharacterType;
 using mojom::ConversationTurn;
 using mojom::ConversationTurnVisibility;
 
-#if BUILDFLAG(ENABLE_PRINT_PREVIEW)
-class PreviewPageTextExtractor {
- public:
-  PreviewPageTextExtractor(base::ReadOnlySharedMemoryRegion pdf_region,
-                           base::OnceCallback<void(std::string)> callback,
-                           uint32_t max_page_content_length,
-                           std::optional<bool> pdf_use_skia_renderer_enabled)
-      : pdf_region_(std::move(pdf_region)),
-        callback_(std::move(callback)),
-        max_page_content_length_(max_page_content_length) {
-    DCHECK(!pdf_to_bitmap_converter_.is_bound());
-    GetPrintingService()->BindPdfToBitmapConverter(
-        pdf_to_bitmap_converter_.BindNewPipeAndPassReceiver());
-    pdf_to_bitmap_converter_.set_disconnect_handler(
-        base::BindOnce(&PreviewPageTextExtractor::BitmapConverterDisconnected,
-                       base::Unretained(this)));
-    if (pdf_use_skia_renderer_enabled.has_value()) {
-      pdf_to_bitmap_converter_->SetUseSkiaRendererPolicy(
-          pdf_use_skia_renderer_enabled.value());
-    }
-  }
-
-  void StartExtract() {
-    pdf_to_bitmap_converter_->GetPdfPageCount(
-        pdf_region_.Duplicate(),
-        base::BindOnce(&PreviewPageTextExtractor::OnGetPageCount,
-                       base::Unretained(this)));
-  }
-
-  void ScheduleNextPageOrComplete() {
-    DCHECK_GT(total_page_count_, 0u);
-    if (current_page_index_ < total_page_count_) {
-      if (current_page_index_) {
-        preview_text_ << "\n";
-      }
-      pdf_to_bitmap_converter_->GetBitmap(
-          pdf_region_.Duplicate(), current_page_index_,
-          base::BindOnce(&PreviewPageTextExtractor::OnGetBitmap,
-                         base::Unretained(this)));
-    } else {
-      std::move(callback_).Run(preview_text_.str());
-    }
-  }
-
-  void OnGetPageCount(std::optional<uint32_t> page_count) {
-    if (!page_count.has_value() || !page_count.value()) {
-      std::move(callback_).Run("");
-      return;
-    }
-    total_page_count_ = page_count.value();
-    ScheduleNextPageOrComplete();
-  }
-
-  void OnGetBitmap(const SkBitmap& bitmap) {
-    if (bitmap.drawsNothing()) {
-      std::move(callback_).Run(preview_text_.str());
-      return;
-    }
-#if BUILDFLAG(ENABLE_TEXT_RECOGNITION)
-    GetOCRText(bitmap,
-               base::BindOnce(&PreviewPageTextExtractor::OnGetTextFromImage,
-                              weak_ptr_factory_.GetWeakPtr()));
-#else
-    std::move(callback_).Run("");
-#endif
-  }
-
-  void BitmapConverterDisconnected() {
-    DLOG(ERROR) << __func__;
-    if (callback_) {
-      std::move(callback_).Run(preview_text_.str());
-    }
-  }
-
-  void OnGetTextFromImage(std::string page_content) {
-    VLOG(4) << "Page index(" << current_page_index_
-            << ") content: " << page_content;
-    preview_text_ << page_content;
-    // Stop processing if we have reached the maximum number of pages or the
-    // maximum length of the content
-    if (current_page_index_ + 1 >= kMaxPreviewPages ||
-        preview_text_.str().length() >= max_page_content_length_) {
-      std::move(callback_).Run(preview_text_.str());
-      return;
-    }
-    ++current_page_index_;
-    ScheduleNextPageOrComplete();
-  }
-
- private:
-  std::stringstream preview_text_;
-  size_t current_page_index_ = 0;
-  size_t total_page_count_ = 0;
-  base::ReadOnlySharedMemoryRegion pdf_region_;
-  base::OnceCallback<void(std::string)> callback_;
-  const uint32_t max_page_content_length_;
-  mojo::Remote<printing::mojom::PdfToBitmapConverter> pdf_to_bitmap_converter_;
-  base::WeakPtrFactory<PreviewPageTextExtractor> weak_ptr_factory_{this};
-};
-#endif
-
 AIChatUIPageHandler::AIChatUIPageHandler(
-    AIChatUI* owner,
     content::WebContents* owner_web_contents,
     content::WebContents* chat_context_web_contents,
     Profile* profile,
     mojo::PendingReceiver<ai_chat::mojom::PageHandler> receiver)
     : content::WebContentsObserver(owner_web_contents),
-      owner_(owner),
       profile_(profile),
       receiver_(this, std::move(receiver)) {
   // Standalone mode means Chat is opened as its own tab in the tab strip and
@@ -206,6 +86,10 @@ AIChatUIPageHandler::AIChatUIPageHandler(
 
   favicon_service_ = FaviconServiceFactory::GetForProfile(
       profile_, ServiceAccessType::EXPLICIT_ACCESS);
+#if BUILDFLAG(ENABLE_PRINT_PREVIEW)
+  print_preview_extractor_ = std::make_unique<PrintPreviewExtractor>(
+      chat_context_web_contents, profile_);
+#endif
 }
 
 AIChatUIPageHandler::~AIChatUIPageHandler() = default;
@@ -579,126 +463,12 @@ void AIChatUIPageHandler::OnGetPremiumStatus(
 }
 
 #if BUILDFLAG(ENABLE_PRINT_PREVIEW)
-void AIChatUIPageHandler::PreviewCleanup() {
-  auto preview_ui_id = owner_->GetPreviewUIId();
-  CHECK(preview_ui_id);
-  PrintPreviewDataService::GetInstance()->RemoveEntry(*preview_ui_id);
-  print_render_frame_->OnPrintPreviewDialogClosed();
-  owner_->DisconnectPrintPrieviewUI();
-}
-
-void AIChatUIPageHandler::OnPreviewReady() {
-  scoped_refptr<base::RefCountedMemory> data;
-  auto preview_ui_id = owner_->GetPreviewUIId();
-  CHECK(preview_ui_id);
-  PrintPreviewDataService::GetInstance()->GetDataEntry(
-      *preview_ui_id, printing::COMPLETE_PREVIEW_DOCUMENT_INDEX, &data);
-  if (!data.get()) {
-    DLOG(ERROR) << "no data from preview id: " << *preview_ui_id;
-    return;
-  }
-  auto pdf_region = base::ReadOnlySharedMemoryRegion::Create(data->size());
-  if (!pdf_region.IsValid()) {
-    DLOG(ERROR) << "Failed allocate memory for PDF file";
-    return;
-  }
-  memcpy(pdf_region.mapping.memory(), data->data(), data->size());
-  std::optional<bool> pdf_use_skia_renderer_enabled;
-  auto* prefs = profile_->GetPrefs();
-  if (prefs &&
-      prefs->IsManagedPreference(::prefs::kPdfUseSkiaRendererEnabled)) {
-    pdf_use_skia_renderer_enabled =
-        prefs->GetBoolean(::prefs::kPdfUseSkiaRendererEnabled);
-  }
-  preview_page_text_extractor_ = std::make_unique<PreviewPageTextExtractor>(
-      std::move(pdf_region.region),
-      base::BindOnce(&AIChatUIPageHandler::OnGetOCRResult,
-                     weak_ptr_factory_.GetWeakPtr()),
-      active_chat_tab_helper_->GetMaxPageContentLength(),
-      pdf_use_skia_renderer_enabled);
-  preview_page_text_extractor_->StartExtract();
-}
-
-void AIChatUIPageHandler::OnGetOCRResult(std::string text) {
-  active_chat_tab_helper_->OnPreviewTextReady(std::move(text));
-  PreviewCleanup();
-}
-
 void AIChatUIPageHandler::MaybeCreatePrintPreview() {
-  const bool print_preview_disabled =
-      profile_->GetPrefs()->GetBoolean(::prefs::kPrintPreviewDisabled);
-  active_chat_tab_helper_->SetPrintPreviewDisabled(print_preview_disabled);
-  if (print_preview_disabled) {
-    return;
-  }
   auto url = active_chat_tab_helper_->web_contents()->GetLastCommittedURL();
   if (!base::Contains(kPrintPreviewRetrievalHosts, url.host_piece())) {
     return;
   }
-  // TODO(darkdh): support pdf preview printing
-  content::RenderFrameHost* rfh =
-      printing::GetFrameToPrint(active_chat_tab_helper_->web_contents());
-  if (rfh) {
-    if (!print_render_frame_.is_bound()) {
-      rfh->GetRemoteAssociatedInterfaces()->GetInterface(&print_render_frame_);
-    }
-
-    print_render_frame_->InitiatePrintPreview(false);
-
-    if (!owner_->IsBound()) {
-      print_render_frame_->SetPrintPreviewUI(owner_->BindPrintPreviewUI());
-    }
-    auto preview_ui_id = owner_->GetPreviewUIId();
-    if (!preview_ui_id) {
-      owner_->SetPreviewUIId();
-      preview_ui_id = owner_->GetPreviewUIId();
-    }
-    CHECK(preview_ui_id);
-
-    // A mininum print setting to avoid PrinterSettingsInvalid
-    auto settings = base::JSONReader::Read(R"({
-   "collate": true,
-   "color": 2,
-   "copies": 1,
-   "deviceName": "Save as PDF",
-   "dpiHorizontal": 300,
-   "dpiVertical": 300,
-   "duplex": 0,
-   "headerFooterEnabled": false,
-   "isFirstRequest": true,
-   "landscape": false,
-   "marginsType": 0,
-   "mediaSize": {
-      "custom_display_name": "Letter",
-      "height_microns": 279400,
-      "imageable_area_bottom_microns": 0,
-      "imageable_area_left_microns": 0,
-      "imageable_area_right_microns": 215900,
-      "imageable_area_top_microns": 279400,
-      "is_default": true,
-      "name": "NA_LETTER",
-      "width_microns": 215900
-   },
-   "pageRange": [  ],
-   "pagesPerSheet": 1,
-   "previewModifiable": true,
-   "printerType": 2,
-   "rasterizePDF": false,
-   "scaleFactor": 100,
-   "scalingType": 0,
-   "shouldPrintBackgrounds": false,
-   "shouldPrintSelectionOnly": false
-  })");
-    CHECK(settings);
-    auto dict = std::move(*settings).TakeDict();
-    dict.Set(printing::kPreviewUIID, preview_ui_id.value());
-    dict.Set(printing::kPreviewRequestID, ++preview_request_id_);
-    dict.Set(printing::kSettingHeaderFooterTitle,
-             active_chat_tab_helper_->web_contents()->GetTitle());
-    dict.Set(printing::kSettingHeaderFooterURL, url.spec());
-    owner_->OnPrintPreviewRequest(preview_request_id_);
-    print_render_frame_->PrintPreview(std::move(dict));
-  }
+  print_preview_extractor_->CreatePrintPreview();
 }
 #endif
